Moving floating local video layout to its own file
